python_code/serverReceptor2/requirements.txt
paho-mqtt<2.0.0
pandas>=2.0.0